q360_project/requirements.txt
# Django and Core
Django==5.1.4
djangorestframework==3.15.2
django-filter==24.3
django-cors-headers==4.5.0

# Database
psycopg2-binary==2.9.10
dj-database-url==2.2.0

# Authentication & JWT
djangorestframework-simplejwt==5.4.0

# Async Tasks
celery==5.4.0
redis==5.2.1
django-celery-beat==2.7.0

# Admin Interface
django-jazzmin==3.0.1

# History & Audit
django-simple-history==3.8.0

# Tree Structure (for departments)
django-mptt==0.16.0

# File Upload & Storage
Pillow==11.0.0

# Report Generation
reportlab==4.2.5
pandas==2.2.3
openpyxl==3.1.5

# Security
python-decouple==3.8
argon2-cffi==23.1.0

# Production Server
gunicorn==23.0.0
whitenoise==6.8.2

# Development Tools
python-dotenv==1.0.1
